project/mapping/requirements.txt
APScheduler==3.0.5
influxdb==2.12.0
docker-py==1.7.2
